+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### Enhancements
 
+* Added Biomedical literature connector for ingest cli.
 * Add `FsspecConnector` to easily integrate any existing `fsspec` filesystem as a connector.
 * Rename `s3_connector.py` to `s3.py` for readability and consistency with the
   rest of the connectors.